% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jp_daily}
\alias{jp_daily}
\title{Daily nation-level dataset of Japanese covid-19 situation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 778 rows and 8 columns.
}
\usage{
jp_daily
}
\description{
More details are available \href{https://github.com/kj-9/covid19jp#japan_daily}{here}.
}
\keyword{datasets}
