% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pref}
\alias{pref}
\title{Master data for Japanese prefectures. More details are available
\href{https://github.com/kj-9/covid19jp#pref}{here}.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 47 rows and 4 columns.
}
\usage{
pref
}
\description{
Master data for Japanese prefectures. More details are available
\href{https://github.com/kj-9/covid19jp#pref}{here}.
}
\keyword{datasets}
