% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pref_daily}
\alias{pref_daily}
\title{Daily prefecture-level dataset of Japanese covid-19 situation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 39104 rows and 11 columns.
}
\usage{
pref_daily
}
\description{
More details are available \href{https://github.com/kj-9/covid19jp#pref_daily}{here}.
}
\keyword{datasets}
