% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pref_weekly}
\alias{pref_weekly}
\title{Weekly prefecture-level dataset of Japanese covid-19 situation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1787 rows and 27 columns.
}
\usage{
pref_weekly
}
\description{
This dataset contains some additional columns like hospital capacity not appear
in `?pref_daily
More details are available \href{https://github.com/kj-9/covid19jp#pref_weekly}{here}.
}
\keyword{datasets}
