% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_data}
\alias{update_data}
\title{Update the covid19jp datasets}
\usage{
update_data(force = FALSE)
}
\arguments{
\item{force}{A logical value. FALSE by default. if set to TRUE, install
updates without prompt question}
}
\value{
No return value, called for updating covid19jp package datasets.
}
\description{
Update the package datasets on the global environment with the
most recent data on the development version.
}
\details{
As the CRAN version is being updated every one-two months, the
development version of the package is being updated on a daily bases. This
function enables to refresh the package dataset to the most up-to-date
data. Changes will be available on the global environment.
}
\examples{
\donttest{
# update with a question prompt
update_data(force = FALSE)

# update without a question prompt
update_data(force = TRUE)
}

}
