% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jp_daily}
\alias{jp_daily}
\title{Japanese daily data of covid-19 situation. More details are available
\href{https://github.com/kj-9/covid19jp#japan_daily}{here}.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 356 rows and 8 columns.
}
\usage{
jp_daily
}
\description{
Japanese daily data of covid-19 situation. More details are available
\href{https://github.com/kj-9/covid19jp#japan_daily}{here}.
}
\keyword{datasets}
